Extract Jophur class
